split instances from models.base
